+++ lib/features/settings/ui/appearance/appearance_screen.dart
@@ -11,7 +11,6 @@ import 'package:storii/features/settings/ui/appearance/font_family.dart';
 import 'package:storii/features/settings/ui/appearance/font_scale.dart';
 import 'package:storii/features/settings/ui/appearance/marquee_speed_tile.dart';
 import 'package:storii/features/settings/ui/appearance/scheme_variant_tile.dart';
-import 'package:storii/features/settings/ui/appearance/scroll_thumb_tile.dart';
 import 'package:storii/features/settings/ui/appearance/system_theme_tile.dart';
 import 'package:storii/features/settings/ui/settings_header.dart';
 
@@ -103,7 +102,6 @@ class AppearanceScreen extends ConsumerWidget {
           const FontScaleTile(),
           const DateTimeFormatTile(),
           const MarqueeSpeedTile(),
-          const ScrollThumbTile(),
           const BinaryBytesTile(),
           const SizedBox(height: 100),
         ],